# docs/credit-engine-openapi.yaml
openapi: 3.0.0
info:
  title: FinFlow Credit Engine API
  description: AI-driven credit scoring and loan offers API
  version: 1.0.0
servers:
  - url: /api/v1
    description: Base API path
paths:
  /score:
    post:
      summary: Calculate credit score
      description: Calculate credit score based on financial metrics
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditScoreRequest'
      responses:
        '200':
          description: Credit score calculated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditScoreResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: Internal server error
  /offers:
    get:
      summary: Get loan offers
      description: Get loan offers based on credit score
      security:
        - bearerAuth: []
      parameters:
        - name: score
          in: query
          description: Credit score (0.0-1.0)
          required: false
          schema:
            type: number
            format: float
            minimum: 0.0
            maximum: 1.0
      responses:
        '200':
          description: Loan offers retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanOffersResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: Internal server error
  /user/data:
    get:
      summary: Export user data
      description: Export all user data (GDPR compliance)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User data exported successfully
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: Internal server error
  /user/data:
    delete:
      summary: Delete user data
      description: Delete all user data (GDPR compliance)
      security:
        - bearerAuth: []
      responses:
        '204':
          description: User data deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: Internal server error
  /health:
    get:
      summary: Health check
      description: Health check endpoint for monitoring
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  service:
                    type: string
                    example: credit-engine
components:
  schemas:
    CreditScoreRequest:
      type: object
      required:
        - income
        - numInvoices
        - avgCashflow
        - delinquencies
      properties:
        income:
          type: number
          format: float
          description: Annual income
          example: 75000
        numInvoices:
          type: integer
          description: Number of invoices in the system
          example: 12
        avgCashflow:
          type: number
          format: float
          description: Average monthly cash flow
          example: 6500
        delinquencies:
          type: integer
          description: Number of payment delinquencies
          example: 0
    CreditScoreResponse:
      type: object
      properties:
        credit_score:
          type: number
          format: float
          description: Credit score (0.0-1.0)
          example: 0.85
        risk_category:
          type: string
          description: Risk category
          enum: [LOW_RISK, MEDIUM_RISK, HIGH_RISK]
          example: LOW_RISK
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the calculation
          example: "2025-05-20T12:34:56"
    LoanOffer:
      type: object
      properties:
        amount:
          type: number
          format: float
          description: Loan amount
          example: 25000
        interest_rate:
          type: number
          format: float
          description: Annual interest rate (%)
          example: 4.5
        term_months:
          type: integer
          description: Loan term in months
          example: 36
        monthly_payment:
          type: number
          format: float
          description: Monthly payment amount
          example: 738.99
    LoanOffersResponse:
      type: object
      properties:
        credit_score:
          type: number
          format: float
          description: Credit score (0.0-1.0)
          example: 0.85
        offers:
          type: array
          items:
            $ref: '#/components/schemas/LoanOffer'
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the offers
          example: "2025-05-20T12:34:56"
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      headers:
        WWW-Authenticate:
          schema:
            type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT